refactor(reports): adopt @branch/lambda-http declarative route table
Replaces the if-chain in handler.ts with a Route[] table (routes.ts) and
one RouteHandler per route (controllers/reports.ts). handler.ts is now
a thin `dispatch(event, { prefix: 'reports', routes })`.

- Local json() and the local async requireAuth() are gone; dispatch
  provides json/OPTIONS/health/404/500 centrally, and
  createAuthGuard(authenticateRequest) replaces the local requireAuth,
  preserving its exact 401 "Authentication required" message.
- Route order preserved from the original if-chain: POST /generate and
  GET /upload-url stay ahead of the /:id pattern they'd otherwise be
  swallowed by (both are 2-segment paths, same as /reports/:id).
- REPORT_ID_ROUTE/REPORT_DOWNLOAD_ROUTE's \d+ constraint is now an
  explicit numeric check in getReport/deleteReport/downloadReport, so a
  non-numeric :id still falls through to the same 404 instead of being
  looked up as a report id.
- report-service.ts is untouched; controllers still parse/validate/
  respond and delegate to it. S3 presigning stays in the controllers,
  matching where it lived in handler.ts.
- Added two route-precedence unit tests (GET /reports/upload-url and
  POST /reports/generate each reach their own controller, not
  /reports/:id or the generic POST /reports controller).
- ROUTES-START/END markers moved into routes.ts around the route array.

No behaviour change: same status codes, messages, validation order,
and S3/TTL values as before.
