+++ apps/backend/lambdas/reports/package.json
@@ -30,6 +30,7 @@
     "@aws-sdk/client-s3": "^3.995.0",
     "@aws-sdk/s3-request-presigner": "^3.995.0",
     "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
+    "@branch/lambda-http": "file:../../../../shared/lambda-http",
     "aws-jwt-verify": "^5.1.1",
     "aws-lambda": "^1.0.7",
     "docx": "^9.5.0",
