+++ apps/backend/lambdas/reports/tsconfig.json
@@ -10,6 +10,6 @@
     "outDir": "dist",
     "sourceMap": true
   },
-  "include": ["*.ts"],
+  "include": ["*.ts", "controllers/**/*.ts"],
   "exclude": ["node_modules", "dist", "dev-server.ts", "swagger-utils.ts"]
 }
